feat: add answers for Part 4, up to Exercise 4.12
Added a test that verifies that if the title or
url properties are missing from the request data,
the backend responds to the request with the status
code 400 Bad Request.

Also, made changes to the HTTP POST route handler

accordingly so that it passes the test.

Exercises 4.8 - 4.14 are stored in the directory

part4/bloglist-backend.