feat: add answers for Part 4, up to Exercise 4.19
Modified adding new blogs so that it is only possible
if a valid token is sent with the HTTP POST request.
The user identified by the token is designated as
the creator of the blog.

Also, modified tests to check the modified
functionality.

Exercises 4.15 - 4.23 are stored in the directory

part4/bloglist-backend.