feat: add answers for Part 5, up to Exercise 5.9
Addressed the issue of the user's name disappearing

after clicking the like button.

It was because the response the server returns after

the PUT request only contains the userId.

I resolved the problem by using the previous user

property from the entry before update, since the user

is assumed to be identical before and after the update.

Exercises 5.5 - 5.12 are stored in the directory

part5/bloglist-frontend.