feat: add answers for Part 6, up to Exercise 6.11
Installed Redux Toolkit for the project. Moved the

Redux store creation into the file store.js and used

Redux Toolkit's configureStore to create the store.

Changed the definition of the filter reducer and

action creators to use the Redux Toolkit's createSlice

function. Also, changed the definition of the anecdote

reducer and action creators to use the Redux Toolkit's

createSlice function.

Exercises 6.9 - 6.13 are stored in the directory

part6/redux-anecdotes.