feat: add answers for Part 6, up to Exercise 6.24
Implemented the error handling for the insertion.

That is, if the input fails to be more at least 5

characters long, the server sends the error message,

and the React Query onError callback and the

notification reducer handles the error message.

Also, the notification works when a new anecdote is

created or an anecdote is voted on. Lastly, the

notification is displayed for the default value of

five seconds, with debouncing also working for new

influx of notifications.

Exercises 6.23 - 6.24 are stored in the directory

part6/query-anecdotes.