feat: refactor the backend for Comment document
Instead of simply storing comments in an array in a

Blog document, a separate Comment document is created.

Hence, when a comment is added to a blog, both

document objects and User document object can refer

to one another. This is much better for scaling the

app in the future.