feat: add answers for Part 9, up to Exercise 9.18
Make it possible to add new diary entries from the

frontend. In addition, both the backend and the

frontend were edited, so that the diaries can be

fetched with or without commentaries with a separate

endpoint in the backend and a toggle button in the

frontend.

Exercises 9.17 - 9.20 are stored in the directory

part9/flight-diary-frontend and part9/flight-diary.