feat: add answers for Part 9, up to Exercise 9.21
Created an endpoint /api/patients/:id to the backend

that returns all of the patient information for one

patient, including the array of patient entries that

is still empty for all the patients. A new interface

is Entry is introduced, and a new property entries

for Patient. All other functionalities are modified

to incorporate the new type and property.

Exercises 9.21 - 9.22 are stored in the directory

part9/patientor_backend and part9/patientor_frontend.