feat: add answers for Part 9, up to Exercise 9.22
Created a page for showing a patient's full information

in the frontend. The user is able to access a patient's

information by clicking the patient's name.

Also, added Morgan middleware to the backend, so that

the backend logs are more informative.

Exercises 9.21 - 9.22 are stored in the directory

part9/patientor_backend and part9/patientor_frontend.