feat: add answers for Part 9, up to Exercise 9.24
Extended a patient's page in the frontend to list

the date, description, and diagnoseCodes of the

patient's entries. Also, refactored the entire app, so

that Zod schemas and actual types are separated,

minimizing the risk of circular imports.

Exercises 9.23 - 9.30 are stored in the directory

part9/patientor_backend and part9/patientor_frontend.