feat: add answers for Part 9, up to Exercise 9.27
Added endpoint /api/patients/:id/entries to the backend,

so that an entry for a patient can be added.

In order to check that required fields are in correct

formats, schemas from Zod are used as parsers in the

middleware.

Exercises 9.23 - 9.30 are stored in the directory

part9/patientor_backend and part9/patientor_frontend.