fix: refactor backend to integrate with frontend
Refactored the backend so that the error handling

in both backend and frontend is more consistent.

Also, addressed the issue of enabling frontend to add

a new patient by making entries property optional and

having an empty array as default. Therefore, a new

patient can be created in frontend first, then entries

can be added later.

Exercises 9.23 - 9.30 are stored in the directory

part9/patientor_backend and part9/patientor_frontend.