feat: add answers for Part12, up to Exercise 12.10
Implemented a todo counter that saves the number of

created todos to Redis. Therefore, whenever a request

is sent to add a todo, the counter is incremented by one.

Also, if a GET /statistics request is made, the current

number of todos is retrieved from Mongo DB and returned.

Exercise 12.10 is stored in the directory

part12-containers-applications/todo-app/todo-backend.